llama3/OwOGPT/backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==2.6.4
pydantic-settings==2.2.1
SQLAlchemy==2.0.29
paho-mqtt==2.0.0
python-dotenv==1.0.1
openai==1.35.7
ollama==0.3.0
aiohttp==3.9.5
orjson==3.10.3
typing-extensions==4.11.0

